+++ ChangeLog.md
@@ -1,5 +1,10 @@
+<<<<<<< HEAD
 # v0.5 (?)
 * Now can handle videos that do not contain the number of frames in the meta-data
+=======
+# v0.4.1 (?)
+* Added `--metric dm-preview` for debugging of the display model
+>>>>>>> main
 
 # v0.4 (19/01/2024)
 * A new calibration with a small improvement in performance
@@ -10,6 +15,7 @@
 * Fixed distogram generation for images
 * "--quiet" flag now ensures that no warning messages end up in the stdout
 * CSF updated to the latest castleCSF fit
+* added 'luminance' color space to handle luminance-only data
 
 # v0.3 (05/09/2023)
 * Added value range checks when the metric is running on HDR data (to avoid passing relative values)